Pretty sure I've fixed balls not colliding (#7)
The physics engine cancels collisions if the two balls that collide are
trying to follow one another. Sometimes the follow-related properties are
set incorrectly on balls, which causes balls to unexpectedly pass through
each other.

This was probably caused by snake head reversal - when a snake is cut in
half, its head can swap ends. This didn't update the follow- properties,
leaving an invalid `.following` on the head, which cancelled some
collisions.

Fixed by clearing the follow- properties on every game update. Not the
tidiest solution, but it should work.